PlayChoice-10 key ROMs, registers, and IO
Dropped the second key file (counter out, which is dynamically generated
based on an address pin) and reduced the expected size of the first key
file from 16 to 9. Added methods to handle reading and writing
PlayChoice-10 registers by the Z80. Because no Z80 code exists in the
processor directory yet, these will go unused for now.