VS. System auto-switch Zapper
Enabled automatically switching to the Zapper when loading a VS. System
game whose manifest is configured to use it. It is unknown at what point
the crashing bug preventing this change disappeared. Note that the VS.
Zapper's light sense is still not working right (the Famicom/NES one
works correctly).