nall::XML::Node fix
In higan v093, changes to nall rendered the XML half of nall/markup/node
unusable. The string that holds node names was receiving data but not
recording the new size because a method ("reserve") was being called
when it should be another method ("resize") that calls the former. This
fix has no practical purpose in the BML-dominated scene and may be
removed when higan v095 is released because nall::XML::Node is
deprecated.