Add locate()
Imported the locate() method from tomoko to nSide-e. This method will
search for a file in the program's install directory first, then fall
back on a specified path, such as the library path, if the former is not
found. This enables one to place the System directories ("Super
Famicom.sys", et al.) in the library path instead of the program's path
if desired.