auto_purify MCC/Satellaview/Sufami Turbo modernization
ananke, the base for auto_purify, contains output for versions of bsnes
from v091 and earlier in its heuristics code, as well as an erroneous
apostrophe in its Sufami Turbo output. These are now fixed.
With one exception: bsnes still takes a "psram" node when loading an MCC
game, which was meant to be changed to "ram" but never was. "psram" is
an artifact from v091 and earlier, and should be changed.