phoenix::Group
A slightly cleaner fix for the nall::group problem. This time, the code
for it is no longer in the nall namespace, and it is renamed to "Group"
(note the uppercase G). Also, the previous commit did not change every
file in phoenix, just the .hpp files, so this one changes the .cpp files
as well.