Import multiple features into nSide-t
In preparation for the discontinuation of nSide-e, some of its features
were transferred over to nSide-t, including:
1. The System menu is renamed to the active console's name upon loading
media.
2. The Fast Forward hotkey is now available.
3. Hotkeys can be bound to joypad buttons.
4. The Display Emulation shader is available, but not other video
shaders.
The first 3 additions are also features in the private beta of higan
v094r42. When higan v095 is released, nSide-t's code for these features
will be modified to match higan's exactly.