Input device bug fixes
When one closes higan/tomoko, the bindings for all device mappings that
had spaces in their names, such as "Port 1 - A", were lost because BML
does not allow spaces in node names. A sanity check that was applied to
console names, port names, and device names was not being applied to
mapping names. This is not fixed in nSide-t. Also, tomoko has a
hard-coded limit of 2 ports in the System menu because of the hiro
convention of defining menu options at compile-time. nSide-t now
increases it to 4, but this is not a permanent fix.