GBA experimental cheat support
Added cheat support to the GBA emulator. However, unlike the other
consoles, it behaves weird here; A cheat code's data and comparison
parts must be as wide as the read operation being performed on the
address, or it will not work. If the address is read using a 32-bit
read, then the data portion of the cheat must be 8 digits long. For
example, in Mario & Luigi: Superstar Saga, to increase Luigi's jump
height from 1 block to 3 blocks, using "02007208/5E+02007209/07" will
not work, but "02007208/075E" will (16-bit read). This is subject to
change in the future if byuu adds GBA cheat code support to higan.