GBA missing SHA256 in cheats fix
The lack of an Emulator::Interface function in nSide-gba was causing it
to generate cheats.bml files with blank sha256 fields. This wouldn't
really affect anything because of the lack of Game Boy Advance cheats in
the cheat database, but it should still be fixed for consistency and for
future developments.