MCC markup compromise
Judging by differences between snespurify and ananke, the "psram" node
in MCC documents was intended to be changed into a 2nd "ram" node for
higan v092. However, that change was never made in bsnes, and hope seems
grim for that change to make it into higan v095. So I'm making the
change in nSide-sfc and adding a fallback to maintain compatibility with
higan v095.