icarus Famicom expansion
nSide's version of icarus needs a new name after this addition: the
Famicom database already used by auto_purify is now available for use,
and it contains new heuristics for the VS. System as well.

Note: heuristics for the VS. System are unreliable, much like Game Boy
Advance heuristics. Not all VS. System games have $4016 and $4017
reversed compared to the Famicom (Super Mario Bros. is one example), and
the iNES header doesn't store DIP switch info, so it will need to be
added by hand if not found in the database.