laevateinn Modernization: Multiple Windows
Fixed errors in the About, CPU, CPU Registers, Memory, SMP, SMP
Registers, and VRAM windows. Also, in the Properties window, where it
displays the size of each BG plane, replaced the letter x with the
multiplication symbol × (U+00D7) and added "x" and "y" suffixes to
indicate which dimension is which.