star-rod VRAM Window Palette Selector
Added a palette selection box to the VRAM window. This will allow
viewing the tiles with the palettes that are available to them in the
display. For example, in 2BPP mode, 4 layers × 8 palettes per layer are
available, in 4BPP mode, 8 background palettes + 8 sprite palettes are
available, and in 8BPP mode, only a single palette is available.