Upgrade SFC, GB, GBA to higan v096, remove nSide-e and auto_purify
Imported all changes from higan and icarus v096 into nSide and cart-pal,
excluding the Famicom. In addition, removed nSide-e (ethos) and
auto_purify (ananke). tomoko has almost caught up with ethos in features
(only missing DIP switches, timing analysis, and the ability to assign
multiple simultaneous keys to a hotkey function), so nSide-e is
redundant and ancient.

This commit excludes the Famicom emulator, nSide-fc. Because of all that
has been changed compared to bnes, a direct substitution is not an
option. It will be restored in a future commit.