Upograde FC to higan v096, nall programpath() bugfix
Manually updated nSide-fc to comply with bsnes's code style, and altered
the manifest formats of the Famicom, PlayChoice-10, and VS. System. Also
included a bugfix for nall's programpath() function in Windows, which
would get thrown off in icarus when kaijuu launches higan.

For the VS. System, games including chips on both sides no longer have a
wrapper node ("cartridge") for each side. Both sides are now stored
directly in the root of the document, and the node name is "side"
instead of "vs".