Convert star-rod from phoenix to hiro
Convert all phoenix function calls in star-rod to hiro function calls,
and moved widget properties and relationship from .cpp files into .hpp
files in accordance with hiro's best practices.

In the process, discovered a bug in hiro::Keyboard that affected
Windows, GTK, and Qt on a target-by-target basis, and pushed a fix for
each one. Only the Windows target fix is tested.