Cartridge::load() <-> System::load() swap
Changed the calling order of Cartridge::load() and System::load() for
all 4 emulators so that System::load() calls Cartridge::load() instead
of the other way around. The Super Game Boy does not work in this
version.

This is a change made by development versions of higan.
Note: Only code structural changes and bug fixes will be imported from
higan's betas.