Bitwise Operators Considered Harmful
Changed most usage of bitwise operators (&, |, <<, >>) to use the bit()
or bits() methods of the Integer and Natural classes, severely dropping
emulation speed and causing collateral damage in the form of explicit
typecasts to cater to viewers who do not know the meanings of the bit
operators.

The performance profile will not compile in this version.

This is a change made by development versions of higan.
Note: Only code structural changes and bug fixes will be imported from
higan's betas.