VS. System refactor, part 2
Reorganized a bunch of code relating to the VS. System, and possibly
fixed a bug preventing execution of games larger than 64 KB.

Right now, VS. Super Mario Bros. will not run without using a trick in
the manifest: Put the game's info in the first "side" node, and add a
second, empty "side" node to force the game to run on the Main side.
Games default to the Sub side when they have only 1 "side" node, and the
shared RAM is mapped to the Main side by default.