Preliminary NES-SFEXPROM support
Added preliminary support for NES-SFEXPROM, a board with a DRM chip that
patches the Bases Loaded 1.0 PRG ROM so that, in every 32 KB bank, byte
$0180 is changed from 0x60 to 0x05. By disassembling the ROM in a
debugger, it is believed that the NES-LD-0-EXP chip places a register at
$7000 and activates the patch when d5 is set.

The included manifest in cart-pal's database uses the sha256 of the
non-patched ROM, which is likely not distributed anywhere or easy to
acquire from a copier; it was calculated by simply editing each of the 8
locations where a byte is replaced. The pre-patched ROM's manifest is
included in a comment.