Famicom unrecognized board warning, Emulator::Interface::notify()
Added support for NAMCOT-3303 and NAMCOT-3304, and added a warning on
unrecognized board IDs that relies on the notify() function in
Emulator::Interface. nSide will still crash, but it will now display the
notification first when loading a Famicom game with an unrecognized
board ID.

The notify() function in nSide-t's interface binding will now display a
notification in a message window.