Interface::deviceChanged()
Added a function to Emulator::Interface that report a device change not
caused by the user to the UI. This allows connecting the NES Four Score
to only take 1 step instead of 2. When disconnecting a Four Score,
nSide-fc will connect "None" to whichever port was not manually switched
out (if the user used Port 1 to disconnect the Four Score, port 2 will
become None, and vice versa).