Complaint about GitHub for Windows
GitHub for Windows is buggy software.

Whenever a file is changed, GitHub for Windows automatically marks it as
ignored, and the user must manually mark it as ready to commit... each
and every time a change is saved.

It is impossible to edit the names and descriptions of previous commits
to correct typos and other mistakes, so anything that slips in is
permanently hardwired into the commit history forever. Commit data such
as file changes can be locked, but commit names and descriptions should
be changeable. Why not?

The previous commit ("Save State Regression Fix") has a completely wrong
name and description all because of the former bug, and it will always
have the wrong name and description because of the latter
defective-by-design flaw.