Famicom and Super Famicom PPU structural changes
Moved some of the Famicom PPU's functions into separate files to mirror
changes in the Super Famicom PPU's functions.

Follow-up to GitHub for Windows complaint: The limitation of not being
able to edit commit names and descriptions may not be GitHub for
Windows's fault, but the fault of the Git protocol. Either way, it fails
to account for human error. Humans make mistakes, and we need to be able
to correct them when we do. Typos in commit data can be fixed by
additional commits, but typos in commit names and descriptions are
forever, and that is defective by design.