Emulator::Video
Merged the common functions of each emulator's video class into
Emulator::Video so that the emulators can feed raw PPU output into it
and not concern themselves with emulating color bleed and interframe
blending. The implementation differs slightly from higan v098r04 in that
Video::refresh is split into 2 separate functions to make it easier to
manage multiple viewports (VS. DualSystem, PlayChoice-10). The refresh()
function itself simply calls the 2 split parts in order, and should only
be used when an emulator needs to draw exactly 1 screen.

Rotation is no longer fully available in the WonderSwan for now, except
at startup. When the emulator decides on an orientation, it is stuck
with that orientation for the entire time the game is loaded.