Famicom APU revisions
Added a list of APU revisions to the Famicom emulator. Information about
the revisions is scarce, and multiple sources contradict each other. For
now, the current assumption is that RP2A03F is bugged and lacks
looped/periodic noise while its preceding revision, RP2A03E, has it.
This could change in the future if the revision differences are ever
clarified.