Game Boy Advance rotation support
Added support for rotation to the Game Boy Advance. Far fewer Game Boy
Advance games use it than WonderSwan games (one example is Dr. Mario &
Puzzle League), but this is mainly a demo to test emulator-agnostic
viewport rotation. Note that this is nSide-original and not supported in
higan v098r05.

Also, the Thread size for all emulators has been increased to 4 times
its size as a test for AMD driver crashes.