Upgrade to v098r11, repair Four Score device change
Rewrote the code that, when the Four Score is connected into Port 1,
automatically connects it to Port 2, and, when any other device is
connected into Port 1, connects None into Port 2 (and vice versa).

In addition, altered the Famicom Controller class to allow only a bool
to be returned by the data1() function, to enforce the fact that only a
single bit can be passed through the expansion port into register $4016.
An expansion port device cannot substitute for Player 1's controller,
but it can substitute for Player 2's controller (strangely enough).

Also altered the SFC Gamepad's code to reorganize the buttons to
compensate for the loss of Interface::Device#order in higan v098r11.