GitHub for Windows bug omits changed files by default
Committed a file omitted by the GitHub for Windows bug I complained
about previously.

When I change a file, do I want it left out of the changes to commit?
No! So I have to manually check each file's checkbox before committing
it to the repository. There is no option in the settings to change this
behavior, so GitHub for Windows is defective.