R6502 split instructions.cpp into opcode_*.cpp, Balanced SMP regression fix
To mimic the structure of the R65816 code, the R6502 instructions are
now split across multiple files grouped by Read, Write,
Read-Modify-Write, Program Counter, Miscellaneous, and Illegal.

Also fixed a regression in the Super Famicom Balanced profile SMP. It
was communicating with blargg's DSP with the expectation that it was
byuu's Accuracy DSP, so it relied on the Scheduler, which blargg's DSP
does not use.