R6502 SHX, SHY fix
Fixed a long-standing issue with the unofficial SHX and SHY opcodes
(called "SXA" and "SYA" by some people).

The only remaining opcode that does not pass blargg's CPU tests is LXA
#imm (0xAB). But that opcode is unstable and has different results on
different production runs of the 6502, which is why nSide's
implementation uses rand() for it. blargg's test expects the "correct"
behavior of the missing LAX #imm opcode.