Controller ID bug fix, VS. System manifest format change
Implemented a suggestion from byuu to fix the controller ID bug
(mismatch between indexes in the port and global IDs). The fix
completely drops the concept of global device IDs and sticks exclusively
with device port indexes. This fix may change when byuu implements it in
higan, so this commit is just a pre-emptive fix.

The VS. System manifest format now requires manifests to use the string
"zapper" instead of "beamgun" to specify the VS. Zapper. It is likely
that no game using the VS. Zapper was ever released in VS. System format
in Japan.