Gamepad with Mic
Added support for a Famicom gamepad with a mic. This is implemented by
ANDing the user input with a random value to simulate the chaotic state
of the input that THE HYRULE FANTASY: ゼルダの伝説1 (Zelda no Densetsu 1)
seems to require. If the mic input is forced to 1, nothing happens to
the Pols Voices, so it must randomly alternate between 1 and 0.

The Gamepad with Mic is a separate controller. You will need to assign
inputs for it in order to use it.