True Hori 4-Players Adaptor support; rename old to "JoyPair"
The Hori 4-Players Adaptor is strange by the standards of Famicom
multitaps: it has 4 controller ports, whereas other multitaps only have
2. Anyone who knows the expansion port's pinout knows it is impossible
to feed data into $4016.d0, making it hard to conceive why this adapter
has a port for the first player.

Turns out, that device uses a protocol similar to the Four Score, except
that it feeds the data into d1 of each register and swaps the
signatures. It does not touch $4016.d0 or $4017.d0 at all.

The previous implementation of the 4-Players Adaptor has been renamed to
"JoyPair" to make room for this new implementation. Note that far fewer
Famicom games support this adapter than games that support the other
multitaps, including only some of the Kunio-kun games (such as Downtown
- Nekketsu Koushin Kyoku - Soreyuke Daiundoukai). For most practical
purposes, it should not be necessary to use the 4-Players Adaptor when
the JoyPair is sufficient (all 4-player games support the latter).
However, homebrew can potentially accomplish 6 players on a traditional
Famicom or 8 players on an AV Famicom + Four Score.