Update to v100
Changed the balanced profile's handling of mosaic to match the accuracy
profile, finally aligning both profiles' PPU member names. Now, star-rod
can be compiled with the accuracy profile or balanced profile.
Obviously, the accuracy version is best used for debugging hacks and
homebrew. However, the balanced version exists at all for the purpose of
debugging breaking changes to the balanced PPU.

Updated the README to reflect the current state of nSide's code
structure and requirements.