UI viewport aspect ratio further adjustment
Further revised the UI's aspect ratio formula.

Now, the window will only stretch if the display exceeds 256 * scale,
and only to the amount needed. It will not arbitrarily stick to 256 *
8.0 / 7.0 * scale.

Also, the information properties "canvasWidth" and "canvasHeight" have
been dropped, reverting back to higan's definitions of "width" and
"height". The VS. DualSystem and PlayChoice-10 Dual Screen model now set
information.width and information.height to 256×120 and 128×240
respectively for the purpose of Medium (×3) scaling.