Hide nSide logo when using program arguments
Created a new nSide logo to be used in the same places where the higan
logo is used. Then, configured the UI to hide the logo when loading a
game through the command line. Though, it will still show if the
"Unload" option is picked, or if a console is selected from the Library.

After much criticism on Twitter, byuu has considered removing the higan
logo from the splash screen, so nSide may revert on this as well.