Atari 2600 timing adjustment
Changed the way the Atari 2600's clock rates are initialized.

Even though the CPU and PIA receive pre-divided clocks that are passed
through the TIA, writing the clock rates this way causes rounding errors
that make TIA sprites drift to the left every few seconds. All 3 chips
need to share the same clock and use the step() function for clock
division in order to avoid rounding errors.

Also fixed an error in the CXM1P register (collisions between missile 1
and each player).

As a side note, the nSide logo no longer appears in the presentation
viewport when a game is not loaded. This mirrors a similar change in
higan. However, the code for it has been commented out rather than
removed.