PlayChoice-10 Video Circuit timings
In the wake of beginning work on the official Z80, added timing code to
the PlayChoice-10 Video Circuit to change it from a frame-based renderer
to a scanline-based renderer. Changing it to a pixel-based renderer
would be as simple as moving a single line and changing it, but it was
not done in order to keep a little speed.

Also cleaned up the Famicom's system.load() function and moved
arcade-specific system manifest loading to the respective machines'
load() functions.