Atari 2600 color generation adjustment
According to AtariAge's TIA schematics, there is a circuit that delays
the chroma by a specified amount when the hue is 2 to 15. Hue 1 does not
use the delay circuit at all, so it must be equal to NTSC colorburst
(180°).

There are 2 different formulas for phase delays across Atari's field
service manuals: one that makes hues 1 and 15 identical (360° / 14), and
one that makes hue 15 equal to 2/3rds the distance to hue 2 from hue 1
(360° / (14 - 2 / 3)).