MOS6502 Decimal mode support
Added decimal mode to the MOS6502 processor, and made the Famicom
capable of choosing not to allow its usage.

Because there is at least 1 emulator that uses a non-Ricoh variant of
this chip, it is no longer accurate to call it "R6502", so it has been
renamed. I used "MOS" instead of "M" to distinguish it from Motorola,
which was in the CPU business for longer and does not have an acronym
for a name.