Remove loki, add tomoko as reference UI
byuu has completely discontinued loki, thinking he may need to extend
tomoko instead of writing a new GUI. As such, tomoko has been imported
into nSide's repository to use as a reference for changelogs. It cannot
actually be used as a UI for nSide, so compiling it will raise an error.
In addition, star-rod's days could be numbered.

In the future, nSide's own UI may be extended further than before, as a
result of no longer being tied to tomoko.