Multiple inputs support
Allowed binding of multiple inputs to a single mapping, such as
Alt+Enter for fullscreen or Ctrl+R for reset. Game controls use OR
logic, so any of the inputs will activate a game control, and hotkeys
use AND logic, so all of the inputs must be pressed together to activate
a hotkey.