PlayChoice-10 connect Z80 to video circuit
Connected the PlayChoice-10's Z80 CPU to the video circuit, and made
some fixes to the code. It still does not fully boot (it hangs after
"WAIT 10 SEC." and before displaying the PlayChoice-10 logo), but it
does successfully change the video circuit's display during the few
seconds it operates, and Achten's unofficial BIOS works correctly except
for unimplemented cartridge switching.

A PlayChoice-10 Panel device was added to the configuration settings,
and the Channel Select, Enter, Reset, and Service Buttons will now work.