Z80 NMI fix
NMIs on the Z80 were broken; normally, they behave identically to
interrupt mode 1, but higan had them read the r.im variable, which meant
sometimes using mode 0 or 2 instead. This was the cause of the
PlayChoice-10's graphical corruption and inability to use the menu. This
has now been fixed, the menu works, and game instructions can now be
viewed. However, the PlayChoice-10 still has features left to emulate,
such as the CPU (2A03) stopping function, and possibly a way to load
games into the other 9 slots.

This commit includes the skeleton of a PC Engine emulator that byuu has
started work on. As always, there is no guarantee that this emulator
will be developed to a playable state.