Move Reset to hardware ports
Created new "Hardware/Controls" devices for every console that has a
Reset button, and removed the "Soft Reset" hotkey.

This means that it is no longer possible to require 2 or more
simultaneous inputs for the Reset button, as it is a game control
instead of a hotkey. Use the Power Cycle hotkey if you want to require
pressing Ctrl for it.

The VS. System and PlayChoice-10 arcade panels are now considered
hardware port devices. If you mapped inputs to them, please remap them
again.