FamicomBox preliminary support
Fixed bugs preventing the FamicomBox from working, and enabled
FamicomBox support. The FamicomBox is treated as an arcade machine just
like the VS. System and PlayChoice-10. However, almost all of its games
are identical to their Famicom equivalents; ProWres (Pro Wrestling) is
the only game known to have a revision made specifically for the
FamicomBox, as it was otherwise only released for the Famicom Disk
System, and its international versions changed the title screen.
FamicomBox games use the ".fcb" extension, so a Famicom game can simply
be renamed to that extension to work, but only if it is part of the
firmware's internal database or has a valid footer at the very end of
the ROM.

There are still many emulation issues, however. Only a single game can
be loaded at a time, so they are hard to debug because of the lack of a
way to fill the other 14 slots.