Update to v102
Changed the "Show Pre-Alpha Cores" option to a set of radio options
named "Development State". This setting distinguishes between "full"
cores, "alpha" cores, and "pre-alpha" cores. A "full" core runs at least
80% of the library with no severe bugs, an "alpha" core runs at least
25% of the library and can lack a major but not game-breaking feature
such as sound emulation, and a "pre-alpha" core can run nearly no games
to completion and is missing many features. The Atari 2600 and SG-1000
are examples of the 3rd type.

Updated the README to reflect the current state of nSide's code
structure and requirements.