More advanced viewport size calculation
Changed the viewport size calculation formula so that the width won't
always be a multiple of 384 pixels even when the game is not PAL. Now,
it will consider the emulator interface's reported video width to
determine whether that width is necessary. It will also consider higan
v102r01's new width of 326 pixels and height of 242 scanlines to
accomodate the PC Engine's buffer.