PlayChoice-10 and FamicomBox multi-cartridge support
Changed the PlayChoice-10 and FamicomBox to support multiple cartridges.
When loading a multi-cartridge arcade machine, it will prompt for each
cartridge in order, but a single "Cancel" tells the system to not load
any more cartridges.

A bug in the FamicomBox makes full cartridge swapping crash, so
cartridge swapping is currently limited in that only the first cartridge
can have an MMC1 or anything that needs timing (not that there's much;
anything with IRQs or memory mapped at $6000-$7fff wouldn't work on an
actual FamicomBox).