cart-pal VS. System heuristics fix
Fixed a long-standing issue with cart-pal's VS. System heuristics that
generated a blank manifest for each VS. System game that was not in the
database. It was generating the manifest before, but said manifest was
never saved to the "markup" variable that was intended for it, so the
markup remained blank.