MOS 6502 rand() fix
Removed rand(), a possible source of Linux compilation errors, from the
MOS 6502's LXA #imm opcode. This opcode is unstable and needs to not
work in order to prevent homebrewers relying on it, so rand() has been
replaced with 0x00.

However, the "No More Secrets" document details 2 use cases where LXA
will always work: when the argument is 0x00 (lxa #$00), or when the
accumulator was 0xff before executing the opcode. The code has been
modified to account for these 2 safe cases.