cart-pal Famicom fix
Fixed a long-standing bug preventing Famicom ROMs from being imported.

The reason this bug remained since the very beginning of nSide's
development (when bsnes v091 was released) is that I use a completely
unrelated tool to do "in-place" purifications, where a ROM gets
converted to a cartridge folder right where the ROM used to be rather
than moved into the library. I deliberately avoided using cart-pal, my
own fork of icarus, for the sole purpose of controlling where the
purification moves my ROMs. Thus, I never found this bug until today.

The only reason I found it when I did is that I made an amazing
discovery in GitHub for Windows: When typing the phrase "lxa #imm" in a
previous commit, I discovered that GitHub for Windows will show a list
of repository issues immediately after typing the number sign (#). This
is the only way that GitHub for Windows will show a list of issues
without requiring the repository owner to go online.

This bug is a result of overcompensating for the iNES header; cart-pal
was repeating a certain operation to avoid including the iNES header in
the SHA256 hash calculation. The repetition happened in 2 separate
locations in the code, and the result was that the program ROM was
assumed to start 32 bytes in instead of 16 bytes.