MMC5 CHR banking new discoveries
AWJ and loopy discovered the way the MMC5's CHR banking registers
actually work:

In 8×8 sprite mode, $5128-$512B are not used at all; $5120-$5127 are
used for both sprites and backgrounds. In 8×16 sprite mode, $5120-$5127
are for sprites only, while $5128-$512B are for backgrounds only. When
using the PPUDATA register at $2007, whichever set of banking registers
written to last will be used.