WDC65816 disassembler address fix, star-rod general fixes
In processor/wdc65816/disassembler.cpp, the disassembler was reporting
inaccurate addresses for branch and jump opcodes. The high byte was
being added to the low byte and 0x00 was being stored where the high
byte should go. For example, [028000] was being reported as [008002].

This commit also fixes some errors in star-rod, including a crashing bug
when attempting to dump CGRAM to a file and a single byte missing from
the ends of WRAM, APU RAM, VRAM, and OAM dumps. The chances of a new
debugger in higan are grim, so this commit should improve star-rod's
life expectancy a little longer.