Save BMP Screenshot
Added a screenshot function to nSide's normal UI. It saves the video
frame as a BMP image in the cartridge folder, with a maximum limit of
10000 screenshots.

Why BMP, which does not support compression at all? Because nall does
not implement a Deflate compressor, and without that, it would be
impossible to make a PNG function.

If PNG support does come in the future, it will replace the current BMP
screenshot function.