ramus: A Header Library For nall Rejects
Added ramus, a new header library for general-purpose functions that
don't belong in nall. ramus includes an MD5 hasher, a Shift JIS to UTF-8
converter, and a Deflate encoder.

The Deflate encoder is used by the PNG encoder to take PNG screenshots,
so the BMP screenshot function has been removed.