star-rod: Remove CPU/SMP Debugger right column
In laevateinn, the original source of star-rod, the CPU and SMP
debuggers did not have the right column displaying the current state of
each processor's flags. In star-rod, this information is inaccurate for
all instructions not located at the program counter, and it is also
redundant to to button at the bottom of each debugger window, so it has
been removed.