SPC700 ORC, AND carry flag fix
Fixed a logic error, which byuu just discovered, in the SPC700 that has
existed for many years that affected the way the instructions ORC and
AND (0x0A, 0x2A, 0x4A, 0x6A) set the carry flag.

Part of higan v102r26 is included in this commit, but the SPC700 has
been omitted because its rewrite is a work in progress.