Move APU RAM from SMP to DSP
As Super Famicom.sys/manifest.bml and
https://wiki.superfamicom.org/snes/files/snes_schematic_color.png show,
the APU RAM is connected directly to the DSP, making it so the SMP needs
to access the DSP to use it. higan is inaccurate to put this memory
directly in the SMP class, so it has been moved to the DSP class in
nSide.

A compilation error fix in the MOS6502 is also included.