Mega Drive 8-bit cartridge RAM support (Temporary)
Added support for 8-bit cartridge RAM, which only occupies odd
addresses, and updated cart-pal to generate manifests that detect the
RAM depth.

This is a temporary change; byuu is aware of the issue and plans to
update higan to also support 8-bit RAM, possibly as early as higan
v103r01. This commit exists because the longer 8-bit RAM is mistaken for
16-bit RAM, the more save RAMs will be destroyed when the emulator
eventually supports 8-bit RAM.