Mega Drive cheat code support
Added support for cheat codes to the Mega Drive.

Note that, just like with the Game Boy Advance, a Mega Drive cheat
code's data width must match the CPU access width, so a 16-bit read
requires 4 digits in the data field and an 8-bit read requires 2 digits.
The M68k splits a 32-bit read into 2 16-bit reads, so a pair of cheat
codes with 4 digits each must be used in that case.