Mega Drive: Default to "Auto" region when using command line
When launching a game with the command line (which requires kaijuu on
Windows), the Mega Drive's region was set to the empty string, skipping
the "Auto" logic that normally reads the region from the manifest and
forcing NTSC-J even if a manifest exists and specifies a different
region. This bug has been fixed.