Master System: VDP background color/priority fix
Previously, higan treated parts of the background where color 0 is
selected as not existing at all and using colors 0-15 in those
locations. But this behavior caused illiteracy in Alex Kidd: High-Tech
World (あんみつ姫); the text area was drawn with a light yellow background,
same as the text itself.

The actual behavior is that background color 0 comes from the same
palette as colors 1-15, with the only difference being that its priority
is forced to 0. The actual backdrop color exists only in the leftmost
column when it is clipped (and overscan areas not rendered by higan).