Mega Drive: mouse peripheral
Added support for the Mega Drive's mouse.

North America got a different variant of the mouse than the other
regions; the clickable trackball was replaced with a 3rd mouse button,
and a 4th button is added to represent Start. nSide's implementation
here represents North America's Mega Mouse, but it is compatible with
the Sega Mouse released elsewhere (the middle mouse button is equivalent
to the clickable trackball).