Mega Drive: Lock-On
Added support for Lock-On cartridges to the Mega Drive emulator. When
launching Sonic & Knuckles, the emulator will prompt for another
cartridge. Cartridges Locked-On this way will pass $200000-$3fffff to
the base cartridge, and writing to $a130f1 affects both cartridges
simultaneously. When the base cartridge's RAM enable flag is turned on,
an internal UPMEM replaces $300000-$3fffff in order to patch Sonic The
Hedgehog 2 into Knuckles in Sonic 2.

A file named "upmem.rom" needs to be put into the Sonic & Knuckles
cartridge folder. If it is absent, the main program ROM will be chopped
into 7/8th of its size and be unusable.