Locale support
Added support for locale customization. For now, the included locales
are en-US and en-GB (which will be used if the system locale is not
available).

"make install" will automatically install the 2 included locales on
Linux and BSD, but Windows users need to manually copy the "Locales/"
folder from "data/" to the installation directory.

Currently, a change in the locale selection requires a restart to take
effect.