"Next Frame" hotkey
Added a hotkey to advance emulation by one frame at a time. Pressing
this hotkey will start emulation and make it stop as soon as the video
has been refreshed, so it is possible to advance by 1/60th of a second
(or 1/50th in PAL or 1/75th for the WonderSwan).